sync

+++ lib/models/temperature.dart
@@ -1,6 +1,6 @@
 class Temperature {
   DateTime dateTime;
-  int temperature;
+  double temperature;
 
   Temperature({required this.dateTime, required this.temperature});
 }
